requirements.txt
blinker==1.9.0
boto3==1.42.96
botocore==1.42.96
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
fonttools==4.62.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.9
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
reportlab==4.4.10
s3transfer==0.16.1
six==1.17.0
tzdata==2026.2
urllib3==2.6.3
Werkzeug==3.1.8
